👌 `matdyn.x`: improve input validation of blocked keywords
Several input `parameters` keywords in our `CalcJob` classes are blocked, since there
is another `CalcJob` input that enforces a value for the input parameter. An example
here is `q_in_cryst_coord` for `matdyn.x`, which must be set to true in case the user
provides a list of k-points in the `kpoints` input.

Instead of the outdated approach of raising an error in the `prepare_for_submission`
script, we move such validation to the `_validate_inputs` method, which is
assigned as validator of the top-level namespace. This means the user will be warned
that in case they set these inputs, their values will be overriden, instead of the
`CalcJob` excepting because of the raised error.